third_party/WebKit/LayoutTests/svg/animations/svgboolean-animation-1-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test 'to' animation of SVGBoolean.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS feConvolveMatrix.preserveAlpha.animVal is false
PASS feConvolveMatrix.preserveAlpha.baseVal is false
PASS feConvolveMatrix.preserveAlpha.animVal is false
PASS feConvolveMatrix.preserveAlpha.baseVal is false
PASS feConvolveMatrix.preserveAlpha.animVal is true
PASS feConvolveMatrix.preserveAlpha.baseVal is false
PASS feConvolveMatrix.preserveAlpha.animVal is false
PASS feConvolveMatrix.preserveAlpha.baseVal is false
PASS successfullyParsed is true

TEST COMPLETE
